Completely decouple the closure compiler from Web Image builder
There is now no longer a direct dependency from the Web Image builder to
any closure compiler code. The support classes are in separate jars
(new) and the classes are loaded via reflection (as before).
This is required for add the wasm codegen jars to the native-image
driver classpath, because before, it would have pulled in the closure
compiler as well